+++ src/com/bmengine/primitives/Position.java
@@ -2,34 +2,34 @@ package com.bmengine.primitives;
 
 public class Position {
 
-    private int x, y;
+    private double x, y;
 
-    public Position(int x, int y) {
+    public Position(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
 
-    public int getXDistance(int xPosition) {
+    public double getXDistance(double xPosition) {
         return this.x - xPosition;
     }
 
-    public int getYDistance(int yPosition) {
+    public double getYDistance(double yPosition) {
         return this.y - yPosition;
     }
 }
